Refactor function body generators
